Clean up plugins.
