+++ plugins/Cargo.toml
@@ -1,8 +1,10 @@
 [workspace]
 resolver = "3"
-members = ["mocked-tool", "wasm-test"]
+members = ["api-usage", "mocked-tool"]
 
 [workspace.dependencies]
 extism-pdk = "1.4.1"
 serde = { version = "1.0.219", features = ["derive"] }
+starbase_utils = { version = "0.11.13", default-features = false }
 tokio = { version = "1.46.1", features = ["full"] }
+tracing = "0.1.41"
