--- plugins/wasm-test/README.md
@@ -1,5 +1,0 @@
-Build for testing:
-
-```
-cargo build --target wasm32-wasip1 -p proto_wasm_test
-```
